+++ tests/unit/backend/common/entities/test_collection.py
