+++ tests/unit/backend/common/test_auth0_manager.py
