+++ tests/unit/backend/common/utils/test_api_key.py
