+++ tests/unit/backend/common/utils/test_color_conversion_utils.py
@@ -7,7 +7,7 @@ from backend.common.utils.color_conversion_utils import (
     convert_anndata_category_colors_to_cxg_category_colors,
 )
 from backend.common.utils.http_exceptions import ColorFormatException
-from tests.unit.backend.corpora.fixtures.environment_setup import fixture_file_path
+from tests.unit.backend.fixtures.environment_setup import fixture_file_path
 
 
 class TestColorConversionUtils(unittest.TestCase):
