+++ tests/unit/backend/wmg/api/test_v1.py
@@ -5,7 +5,7 @@ from unittest.mock import patch
 from backend.api_server.app import app
 from backend.wmg.api.v1 import find_dim_option_values
 from backend.wmg.data.schemas.cube_schema import cube_non_indexed_dims
-from tests.unit.backend.corpora.fixtures.environment_setup import EnvironmentSetup
+from tests.unit.backend.fixtures.environment_setup import EnvironmentSetup
 from tests.unit.backend.wmg.fixtures.test_primary_filters import (
     test_snapshot_id,
     test_organism_terms,
